+++ main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/previder/previder-go-sdk/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}