+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2024-09
+- Added Virtual Firewall support
+- Seperated CLI into [previder-cli](github.com/previder/previder-cli)
+- Renamed BaseClient to PreviderClient
+- Renamed all implementations to Impl
+
+## 2024-08
+- Added STaaS support
+
 ## 2024-06
 ### Changed
 - Updated to Go 1.22